+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added SESSION_EXPIRES_IN_SECONDS config to myems-api
-- added virtual power plant to database
+- added virtual power plant to database, myems-api and myems-admin
 - 
 ### Changed
 -